diarization-docker/diarization-service/requirements.txt
# Core libraries
dotenv
torchaudio
transformers
soundfile
librosa
omegaconf
wget
tqdm
text-unidecode
ipython
huggingface_hub
grpcio==1.64.1
grpcio-tools==1.64.1
tenacity
requests
pydub
tenacity
# NeMo from GitHub (ASR support)
nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git@main

